sunxi: video: Add video driver for H3/H5/A64 SoC
This patch adds support for hdmi and composite output. It is designed
in the same way as video driver for older Allwinner SoCs.

HDMI: First it checks if monitor is attached. If it is, recommended
timings are read from EDID. After that, DE2, TCON and HDMI are
configured according to this timings.

Composite output doesn't use any check if monitor is attached. This
can be added later.

32MB of RAM is used for framebuffer. This is just enough to support
4K resolution.

SimpleFB is also supported by this driver.

Signed-off-by: Jernej Skrabec <jernej.skrabec@siol.net>